// sourcecode/BTrak.Packages/snova-dynamic-module/dynamic-module/src/app/models/module.ts
export class ModuleModel {
    dynamicModuleId: any;
    dynamicModuleName: any;
    dynamicTabs: any[];
    viewRole: any;
    editRole: any;
    viewRoleName:any;
    editRoleName:any;
    isArchived: any;
    moduleIcon: any;
    base64: any;
    fileName: any;
    timeStamp: any;
    isEdit: any;
    
}

export class  TabModel {
    dynamicTabId: any;
    dynamicTabName: any;
    viewRole: any;
    editRole: any;
    viewRoleName:any;
    editRoleName:any;

}