// sourcecode/BTrak.Packages/snova-dynamic-module/dynamic-module/src/app/models/role-model.ts
export class RoleModel {
  roleId: string;
  roleName: string;
  features: string[];
  featureIds: string[];
  isArchived: boolean;
  featureIdXml: string;
  data: string;
  timeStamp: any;
  isDeveloper: boolean;
  dynamicModuleName: string;
  moduleIcon: string;
  base64: string;

    fileName: string;
  viewRole: string;
  editRole: string;
  
}